pyproject.toml
[project]
name = "moosick"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">3.12, <3.13"
dependencies = [
    "py-cord[voice]>=2.6.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
